chat_service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-socketio==5.11.1
redis==5.0.1
python-jose[cryptography]==3.3.0
requests==2.31.0
python-dotenv==1.0.1
pydantic-settings==2.2.1
httpx
python-multipart
passlib[bcrypt]
sqlalchemy>=2.0.39
alembic>=1.13.1
psycopg2-binary>=2.9.10
asyncpg>=0.29.0 